agregado algunas funciones de alertas y el smoot scroling

+++ src/components/CardFavorites/CardFavorites.jsx
@@ -3,16 +3,41 @@ import { useNavigate } from "react-router-dom";
 import { CardContainer } from "./style";
 import { removeCharacterToFavorite } from "../../redux/character/actions";
 import { useDispatch, useSelector } from "react-redux";
+import Swal from "sweetalert2";
 
 export const CardFavorite = ({ character }) => {
-  const navigate = useNavigate()
+  const navigate = useNavigate();
   const favorites = useSelector((state) => state.character.favorites);
   const dispatch = useDispatch();
- 
-  const handleRemoveFavorites = () => {
-    alert(`se eliminara a ${character.name} de favoritos` )
-    dispatch(removeCharacterToFavorite(character.id));
 
+  const handleRemoveFavorites = () => {
+    Swal.fire({
+      icon: "question",
+      title: "Favorites",
+      html: `<p> Estas seguro de eliminar a ${character.name} de favoritos?</p>`,
+      showDenyButton: true,
+      denyButtonText: "No",
+      denyButtonColor: "blue",
+      confirmButtonText: "Si",
+      confirmButtonColor: "red",
+    }).then((response) => {
+      if (response.isConfirmed) {
+        dispatch(removeCharacterToFavorite(character.id));
+        Swal.fire(
+          "Favorites",
+          `El personaje ${character.name} fue eliminado`,
+          "success"
+        );
+      } else if (response.isDenied) {
+        Swal.fire(
+          "Favorites",
+          `El personaje ${character.name} no fue eliminado`,
+          "error"
+        );
+      } else {
+        Swal.fire("Error", "ocurrio algo inesperado", "error");
+      }
+    });
   };
 
   return (
